chore(plugin): move image helpers to plugin package
