+++ plugins/nemo-data-designer/src/nemo_data_designer_plugin/jobs/create.py
@@ -21,7 +21,7 @@ from nemo_platform_plugin.jobs.api_factory import (
     PlatformJobSpec,
     PlatformJobStep,
 )
-from nmp.common.jobs.image import get_qualified_image
+from nemo_platform_plugin.jobs.image import get_qualified_image
 from pydantic import BaseModel
