+++ plugins/nemo-evaluator/src/nemo_evaluator/jobs/compiler.py
@@ -19,7 +19,7 @@ from nemo_platform_plugin.jobs.constants import (
     DEFAULT_JOB_STORAGE_PATH,
     PERSISTENT_JOB_STORAGE_PATH_ENVVAR,
 )
-from nmp.common.jobs.image import get_qualified_image
+from nemo_platform_plugin.jobs.image import get_qualified_image
 
 EVALUATE_STEP_NAME = "evaluate"
 _RESERVED_SECRET_ENV_NAMES = frozenset({PERSISTENT_JOB_STORAGE_PATH_ENVVAR})
